+++ test/plugin/rack_response_test.rb
@@ -51,19 +51,31 @@ describe Shrine::Plugins::RackResponse do
     assert_equal "application/octet-stream", response[1]["Content-Type"]
   end
 
-  it "returns Content-Disposition with filename metadata" do
+  it "returns Content-Type header from :type" do
+    uploaded_file = @uploader.upload(fakeio(content_type: "text/plain"))
+    response = uploaded_file.to_rack_response(type: "text/plain; charset=utf-8")
+    assert_equal "text/plain; charset=utf-8", response[1]["Content-Type"]
+  end
+
+  it "returns Content-Disposition filename from metadata" do
     uploaded_file = @uploader.upload(fakeio(filename: "plain.txt"))
     response = uploaded_file.to_rack_response
     assert_equal "inline; filename=\"plain.txt\"", response[1]["Content-Disposition"]
   end
 
-  it "returns Content-Disposition with id if filename metadata is missing" do
+  it "returns Content-Disposition filename from :filename" do
+    uploaded_file = @uploader.upload(fakeio(filename: "plain.txt"))
+    response = uploaded_file.to_rack_response(filename: "custom-filename.txt")
+    assert_equal "inline; filename=\"custom-filename.txt\"", response[1]["Content-Disposition"]
+  end
+
+  it "returns Content-Disposition filename with id if metadata is missing" do
     uploaded_file = @uploader.upload(fakeio, location: "foo/bar/baz")
     response = uploaded_file.to_rack_response
     assert_equal "inline; filename=\"baz\"", response[1]["Content-Disposition"]
   end
 
-  it "returns Content-Disposition with disposition applied" do
+  it "returns Content-Disposition disposition from :disposition" do
     uploaded_file = @uploader.upload(fakeio)
     response = uploaded_file.to_rack_response(disposition: "attachment")
     assert_equal "attachment; filename=\"#{uploaded_file.id}\"", response[1]["Content-Disposition"]